refactor(frontend): extract DiffView from SaveDialog
